pyproject.toml
[project]
name = "youtube-nepali"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "audioop-lts; python_version >= '3.13'",
    "google-genai>=1.47.0",
    "google-generativeai>=0.8.5",
    "json5>=0.12.1",
    "moviepy>=2.2.1",
    "pydub>=0.25.1",
    "youtube-transcript-api>=1.2.3",
    "yt-dlp[default]>=2025.10.22",
]

